Move extension methods for combining nodes in CompositeNodeExtensions to a separate namespace
